third_party/blink/web_tests/fast/lists/decimal-leading-zero.html

<ol style="list-style-type: decimal-leading-zero" start="-10">
<li>this list item should be labeled -10</li>
<li>this list item should be labeled -9</li>
<li>this list item should be labeled -8</li>
<li>this list item should be labeled -7</li>
<li>this list item should be labeled -6</li>
<li>this list item should be labeled -5</li>
<li>this list item should be labeled -4</li>
<li>this list item should be labeled -3</li>
<li>this list item should be labeled -2</li>
<li>this list item should be labeled -1</li>
<li>this list item should be labeled 00</li>
<li>this list item should be labeled 01</li>
<li>this list item should be labeled 02</li>
<li>this list item should be labeled 03</li>
<li>this list item should be labeled 04</li>
<li>this list item should be labeled 05</li>
<li>this list item should be labeled 06</li>
<li>this list item should be labeled 07</li>
<li>this list item should be labeled 08</li>
<li>this list item should be labeled 09</li>
<li>this list item should be labeled 10</li>
<li value="100">this list item should be labeled 100</li>
</ol>
<script src="resources/dump-list.js"></script>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        filterListsWithReplacement(document.querySelectorAll("ol"));
    }
</script>
