third_party/blink/web_tests/fast/lists/decimal-leading-zero-expected.txt

-10 this list item should be labeled -10
-9 this list item should be labeled -9
-8 this list item should be labeled -8
-7 this list item should be labeled -7
-6 this list item should be labeled -6
-5 this list item should be labeled -5
-4 this list item should be labeled -4
-3 this list item should be labeled -3
-2 this list item should be labeled -2
-1 this list item should be labeled -1
00 this list item should be labeled 00
01 this list item should be labeled 01
02 this list item should be labeled 02
03 this list item should be labeled 03
04 this list item should be labeled 04
05 this list item should be labeled 05
06 this list item should be labeled 06
07 this list item should be labeled 07
08 this list item should be labeled 08
09 this list item should be labeled 09
10 this list item should be labeled 10
100 this list item should be labeled 100
